Major cleanup: remove personal files, de-dupe code, address PR comments
Removed untracked/personal files:
- workrave productivity tracking (utils/convert_workrave.py, data_fetcher.js, samples/)
- Personal docs (docs/synologygitsetupnotes.md, docs/PR_KNOWLEDGE_BASE.md)
- Broken archive (knowledge-base/archive/ai_conversation.py with syntax errors)
- Empty placeholders (knowledge-base/ip-context/*-extracted.txt)
- Temp diff files (mandates.diff, validate_mandates.diff)

Code de-duplication:
- Created utils/ui_utils.py with shared UI detection functions
- Refactored utils/check_pr_ui_screenshots.py to use shared module
- Refactored utils/compare_ui_screenshots.py to use shared module
- Eliminated duplicate run_git_diff(), is_ui_file(), UI_FILE_PATTERNS, UI_EXCLUDE_PATTERNS

Updated .gitignore:
- Added workrave, CIIRL, docker, psych_notes, mcpo patterns
- Added *.diff pattern
- Organized personal/experimental exclusions

Addresses Copilot review feedback about broken archive and empty files.
