--- docs/PR_KNOWLEDGE_BASE.md
@@ -1,66 +1,0 @@
-# PR: Add Persistent Knowledge Base System
-
-**From:** `feature/knowledge-base-system`
-**To:** `main`
-
-## Title
-Add persistent knowledge base system for AI context and IP documentation
-
-## Description
-
-### Summary
-Adds persistent knowledge base system to enable AI agents (Copilot, Gemini, etc.) to maintain context about Arboreum IP across sessions without re-explaining.
-
-### Key Additions
-
-**Knowledge Base Structure (`knowledge-base/`)**
-- `ip-context/tech-brief.md`: US 19/424,106 technical details (ag-waste bio-refinery, bio-oils)
-- `ip-context/aif-pillars.md`: Arboreum Impact Foundation 85% donation model
-- `ip-context/context.md`: Strategic positioning, outreach history, personal context
-- `QUICK_START.md`: How to use the AI Council system
-- `VSCODE_REMOTE_SETUP.md`: Remote development workflow guide
-
-**AI Conversation System (`tools/ai-evaluator/`)**
-- Multi-AI council for strategic analysis (Gemini, OpenAI, Claude support)
-- Persistent context loading from knowledge base
-- `run_agent.ps1` wrapper for easy execution
-- Requirements: `ai_conversation_requirements.txt`
-
-### Critical Context Preserved
-- **Technology**: Self-powered ag-waste bio-refinery with bio-oil output (SAF, shipping fuel replacement)
-- **Patent**: US 19/424,106 filed 12/17/2025 (credibility anchor)
-- **Phase 2**: Trade secret industrial retrofit IP addressing 60% of emissions
-- **Challenge**: Individual inventor facing institutional gatekeeping
-- **Outreach History**: January 2026 attempts (officials, USDA, Ken Calvert, CSOs) = complete silence
-- **Traffic Validation**: Real qualified visitors (DC, Silicon Valley, Frankfurt, Singapore) - not bots
-- **Accessibility**: User has limited hand use - automation preferred
-
-### Configuration Notes
-- AI Council currently uses Gemini (free tier, working)
-- OpenAI/Claude keys need updating for multi-provider diversity
-- Working preferences documented: execute commands, don't just suggest
-
-### Testing Performed
-✅ Knowledge base loads correctly into AI agent context
-✅ AI Council generates strategic analysis referencing IP
-✅ No re-explanation needed across sessions
-✅ Critical review completed (addresses SEP framing, proof point gaps, timeline reality)
-
-### Next Steps (Future Work)
-- Multi-AI consultation for outreach strategy
-- WIPO GREEN DB entry reconstruction  
-- Contact list generation (climate VCs, patent brokers, journalists)
-- Site conversion optimization (landing page already solid)
-
----
-
-**Note**: This PR does not modify `main` branch site content - only adds backend tooling and persistent documentation.
-
-## To Create PR
-
-```bash
-# Via GitHub CLI (if installed)
-gh pr create --base main --head feature/knowledge-base-system --title "Add persistent knowledge base system for AI context and IP documentation" --body-file PR_KNOWLEDGE_BASE.md
-
-# Or via web: https://github.com/vcaboara/vcaboara.github.io/compare/main...feature/knowledge-base-system
-```
